+++ include/mppic/critics/path_follow_critic.hpp
@@ -16,11 +16,11 @@ public:
   void score(CriticData & data) override;
 
 protected:
-  double max_path_ratio_{0};
+  float max_path_ratio_{0};
   size_t offset_from_furthest_{0};
 
   unsigned int power_{0};
-  double weight_{0};
+  float weight_{0};
 
 };
